Some Test Sprites
READ:

Knight Spritesheet contains 4 way movement (Left can be mirrored to be
right, let me know if it needs to be put on the actual sheet.
It also contains 4 way attack and 4 way ranged (left/right also need to
be mirrored).

Other sheet contains some environment, the ground tiles are in the top
left.